Feature: Scaling Bar HUD
Implemented a cool down to the thruster boost when holding the left shift key. Added a GUI bar to indicate the charge state. The bar depletes as the user holds shift, and refills after the user has let go. If a user has depleted the gauge, the boost effect is disabled.