Feature: Wave System
Added a Wave System based off scriptable objects. This gives me control over how many enemies spawn in each wave, as well as how frequently. There is a small break between each wave. After the final wave is completed, the final wave loops. This can be expanded upon easily to allow different enemies, such as bosses to be added in to waves in the future